exchange_monitor/requirements.txt
requests
pandas
matplotlib
mplfinance
python-dotenv
psycopg2-binary
schedule